order sync logic

+++ com.intime.fashion.data.tmall/Models/SubOrder.cs
@@ -6,8 +6,8 @@ namespace com.intime.fashion.data.tmall.Models
     public partial class SubOrder
     {
         public int Id { get; set; }
-        public decimal TmallOrderId { get; set; }
-        public decimal TmallSubOrderId { get; set; }
+        public long TmallOrderId { get; set; }
+        public long TmallSubOrderId { get; set; }
         public bool LogisticsSynced { get; set; }
         public System.DateTime CreateDate { get; set; }
         public System.DateTime UpdateDate { get; set; }
